TASK: Set up crowdin integration (again)
This updates the crowdin.json file for the 2.3 branch and adds a shell
script to sync sources and translations with crowdin.

That shell script basically contains the code that was in the job on
the Jenkins server as a build step.

A script to install composer is added as well, it can be used by other
scripts, too.